Исправлено: работа с бд, теперь на стартовом экране данные получаются через Flow и Map Рефакторинг

+++ app/src/main/java/com/example/artyomkafood/feature_food/presentation/day/DayViewModel.kt
@@ -2,23 +2,18 @@ package com.example.artyomkafood.feature_food.presentation.day
 
 import android.annotation.SuppressLint
 import android.util.Log
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.artyomkafood.core.basemodel.BaseViewModel
 import com.example.artyomkafood.core.database.Schedule
 import com.example.artyomkafood.core.database.dao.ScheduleMeal
 import com.example.artyomkafood.core.database.entity.MealEntity
-import com.example.artyomkafood.di.viewModelModel
 import com.example.artyomkafood.feature_food.domain.repository.MealRepository
 import com.example.artyomkafood.feature_food.domain.repository.ScheduleRepository
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
 import java.text.SimpleDateFormat
-import java.util.Calendar
+import java.util.*
 
 class DayViewModel(
     private val scheduleRepository: ScheduleRepository,
@@ -30,6 +25,14 @@ class DayViewModel(
 
     private val calendar = Calendar.getInstance()
 
+    fun setExpanded(position: Int) {
+        viewModelScope.launch {
+            data.collect {
+                it[position].expanded = !it[position].expanded
+            }
+        }
+    }
+
     fun getDate(): String =
         simpleDateFormat.format(calendar.timeInMillis)
 
@@ -39,15 +42,16 @@ class DayViewModel(
         getSchedule()
     }
 
-    fun getSchedule()=
+    fun getSchedule() =
         viewModelScope.launch(Dispatchers.IO) {
-            val schedule = scheduleRepository.getSchedule()
-            val meal = scheduleRepository.getMeal(setDate())
-
-            meal.forEach { scheduleMeal ->
-                schedule[scheduleMeal.schedule_id_merge - 1].meal.add(scheduleMeal)
-            }
-            _data.emit(schedule)
+            val schedule = scheduleRepository.getSchedule(setDate())
+            scheduleRepository.getMeal(setDate()).onEach { mealSchedule ->
+                schedule.forEach {
+                    if (mealSchedule[it.name] != null)
+                        it.meal = mealSchedule[it.name]!!
+                }
+                _data.emit(schedule)
+            }.launchIn(viewModelScope)
         }
 
     fun setDate() =
@@ -58,8 +62,18 @@ class DayViewModel(
             set(Calendar.DAY_OF_MONTH, calendar.get(Calendar.DAY_OF_MONTH))
         }.timeInMillis
 
-    fun onSwipeEvent(meal: ScheduleMeal) =
+    fun onSwipeEvent(meal: ScheduleMeal?) =
         viewModelScope.launch(Dispatchers.IO) {
-            mealRepository.delete(MealEntity(meal.meal_volume, setDate(), meal.meal_id))
+            if (meal != null)
+                mealRepository.delete(MealEntity(meal.meal_volume, setDate(), meal.meal_id))
         }
+
+    fun updateMeal(meal: ScheduleMeal?, text: String) {
+        viewModelScope.launch(Dispatchers.IO) {
+            if (meal != null)
+                mealRepository.update(MealEntity(text.toInt(), meal.meal_data, meal.meal_id))
+        }
+    }
+
+
 }
